mqtt_fastapi/requirements.txt
fastapi
uvicorn
paho-mqtt
mysql-connector-python
python-dotenv